Added the possibility to run multipe actions & added tests As of this change you can specify multiple events and actions for a repo by defining a actions hash in tutter.yaml. Like so: ``` projects:   - name: 'JHaals/testing'     access_token: ''     github_site: 'https://github.com'     github_api_endpoint: 'https://api.github.com'
    actions:
      issue:
        - thanks
        - sassy
      issue_comment:
        - sassy

    action_settings:
        whatever_action_specific_setting: 'foo'
        another_setting: false
```

Tutter will look for the specifed actions and rund them in the sequence
specified. If an action where to return anything other than `200 OK` the chain
will stop at the action that broke the chain and only return a response for
that action. With the config above and a hook for an isse the response would
be some thing like:

```
{
  "event":"issue",
  "responses":{
                "thanks":{
                          "status_code":401,
                          "message":"Authorization to JHaals/testing failed, please verify your access token"
                         }
              }
}
```
Some test where also added to check the most common error codes.